barra de busqueda en tabla-servicios funcional y con estilos
